api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openvino==2024.4.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.0.1
python-multipart==0.0.6
aiohttp==3.9.1
pydantic==2.5.0